+++ scripts/main.js
@@ -134,6 +134,13 @@ Hooks.once("ready", () => {
     }
     return ret;
   });
+  Handlebars.registerHelper(MageMagicAddon.ID + "-has-spellslot", function (spellSlots, level, options) {
+    var slots = Object.values(spellSlots)[0];
+    return slots[level] && slots[level].available > 0
+      ? options.fn(this)
+      : options.inverse(this);
+
+  });
 });
 
 Hooks.once("devModeReady", ({ registerPackageDebugFlag }) => {