+++ scripts/utils.js
@@ -1,3 +1,5 @@
+import { MageMagicAddon } from "./MageMagicAddon.js";
+
 const SWNRSkillList = ["Know Magic", "Use Magic"];
 const MTASkillList = ["Gnosis", "Death", "Fate", "Forces", "Life", "Matter", "Mind", "Prime", "Space", "Spirit", "Time"];
 
@@ -24,7 +26,19 @@ export function validateSkillsExist(token, ) {
   return false;
 }
 
-export function isArcanist(actor) {
+export async function isArcanist(actor) {
+  //Check if they have an active form marked as arcanist
+  const activeForm = await actor.getFlag(MageMagicAddon.ID, MageMagicAddon.FLAGS.ACTOR_ACTIVE_FORM);
+  if (activeForm) {
+    const activeFormData = actor.items.contents.find(i => i.id == activeForm);
+    if (activeFormData) {
+      const activeFormCasterType = await activeFormData.getFlag(MageMagicAddon.ID, MageMagicAddon.FLAGS.ITEM_FOCUS_CASTER_TYPE);
+      if (activeFormCasterType && activeFormCasterType == 'arcanist') {
+        return true;
+      }
+    }
+  }
+  // Check if they have the arcanist class
   return (
     actor.items.contents.findIndex(
       (i) =>
@@ -33,7 +47,19 @@ export function isArcanist(actor) {
     ) != -1
   );
 }
-export function isMagister(actor) {
+export async function isMagister(actor) {
+    //Check if they have an active form marked as magister
+  const activeForm = await actor.getFlag(MageMagicAddon.ID, MageMagicAddon.FLAGS.ACTOR_ACTIVE_FORM);
+  if (activeForm) {
+    const activeFormData = actor.items.contents.find(i => i.id == activeForm);
+    if (activeFormData) {
+      const activeFormCasterType = await activeFormData.getFlag(MageMagicAddon.ID, MageMagicAddon.FLAGS.ITEM_FOCUS_CASTER_TYPE);
+      if (activeFormCasterType && activeFormCasterType == 'magister') {
+        return true;
+      }
+    }
+  }
+  // Check if they have the magister class
   return (
     actor.items.contents.findIndex(
       (i) =>
@@ -53,8 +79,8 @@ export function isMtAMage(actor) {
   );
 }
 
-export function isSwNMage(actor) {
-  return isArcanist(actor) || isMagister(actor);
+export async function isSwNMage(actor) {
+  return await isArcanist(actor) || await isMagister(actor);
 }
 
 export function isPsychic(actor) {